--- Client/src/components/reusable-components/icon_text_input_field/icon_text_input_field.tsx
@@ -1,17 +1,0 @@
-import "./icon_text_input_field.css";
-
-interface IconProps {
-  children: any;
-  placeHolder: string;
-  type: string;
-}
-
-export function IconInput(props: IconProps) {
-  const { children, placeHolder, type } = props;
-  return (
-    <div className="wrap">
-      <div className="icon-wrap">{children}</div>
-      <input type={type} placeholder={placeHolder} />
-    </div>
-  );
-}
