+++ Client/src/pages/Home/home.css
@@ -7,6 +7,7 @@
 .container {
   display: grid;
   grid-template-columns: 1fr 1fr;
+  margin-top: 10%;
   margin-left: 25%;
   margin-right: 25%;
 }
